Update visualization.py
fortuneOne() has been modified, now users can search case-insensitively by company's name, or by part of the name. And if the company is not in our list, it will ask user if he or she wants us to search that company online for him

A new empty function named searchOnline() is created.